refactor(engine): split event_translation.go and types.go
Three engine files crossed the file-size trajectory line on this
branch. Fix 6 split runloop_tools.go. This commit does the same
treatment for the other two — by the same rule: documenting a
problem instead of fixing it is the wrong move, and the natural
seams already exist.

event_translation.go (was 707/800 → 577/800):
- fireCliTurnHooks (lines 406-533, ~128 lines) translates raw backend
  events into hook firings. Structurally distinct from the event-
  translation switch above it (translation is pure mapping; hook
  firing is side-effecting and reaches into the extension subsystem).
- Move to event_translation_cli_hooks.go sibling. backend import
  drops out of event_translation.go (no longer needed there).

types.go (was 732/800 → 262/800):
- The cross-cutting types package documents one-concept-per-file as
  its existing pattern (config.go, llm.go, normalized_event.go,
  stream_events.go, timeouts.go, tools.go, + the residual types.go).
  EngineEvent (~290 lines) and RunOptions (~170 lines) are the two
  largest concept blocks in the residual file; promote each to its
  own sibling.
- engine_event.go: the tagged-union EngineEvent struct with its
  comment block. Imports encoding/json (json.RawMessage used for
  Extra fields).
- run_options.go: RunOptions + MessageEndUsage. No external imports
  needed.

Bonus: slash-expand.test.ts (was 596/600 → 542/600 + new
slash-expand-frontmatter.test.ts at 149 lines). The Fix 9 js-yaml
robustness cases pushed slash-expand.test.ts to 667 lines; split
the parseFrontmatter describe block to its own sibling. The fixture
helpers (homedir mock, writeCommand helper) stay in the parent file
since parseFrontmatter has no filesystem dependencies.

File sizes after all splits:
- event_translation.go: 577 (was 707).
- event_translation_cli_hooks.go: 139 (new).
- types.go: 262 (was 732).
- engine_event.go: 307 (new).
- run_options.go: 179 (new).
- slash-expand.test.ts: 542 (was 596 pre-Fix-9).
- slash-expand-frontmatter.test.ts: 149 (new).

Pure mechanical motion. cd engine && go test -race -count=1 ./...
passes; cd desktop && npm test passes (507 tests). make check-file-
sizes OK. cd engine && go test ./internal/types/ -run
TestContractManifest unchanged.
